Implement a reactive state library for propagating FSM to side effects
This is a simple solidjs (createEffect) inspired thing that checks for value equality before dispatching the effect.
This moves a lot of code outside of loop() to setup(), which is more declarative.

Also this ensures we don't contoniously write hw registers.
For example calling setBrightness does a fair bit of calculation.