Add a millisecond count to StateUpdate
Adds the current time of machine since startup in milliseconds as returned by the Arduino function millis().

This is intended to be used for a client to synchronise time with
the server so it can make sense of things like shot timers
which aren't expressed as a straight up count but as timestamps.

This number can also be used to display the uptime of the machine.