Fix setting setpoint in steam mode by changing how the signal is set
Previously the setpoint was flip-flopping by storing a copy of the
previous setpoint before entering steaming.

Now the setpoint is entirely ephemeral and resets to the config value
on transition back to the previous state.

While this UX is slightly less nice in some ways it's less weird and
the state models better.